+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `BoxCoxTransformRegressor`, `TargetModifierRegressor` to `compose`
 - `iter_csv` to `stream`
 - `fetch_restaurants` to `datasets`
+- `GaussianNB` to `naive_bayes`
 
 ## [0.0.2] - 2019-02-13